BAEL-2832: Moved test, and renamed some packages
