+++ java-strings-2/pom.xml
@@ -71,7 +71,7 @@
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
 			<version>${junit.version}</version>
-<!-- 			<scope>test</scope> -->
+ 			<scope>test</scope>
 		</dependency>
 		<dependency>
 			<groupId>org.junit.jupiter</groupId>
