+++ java-strings-2/src/main/java/com/baeldung/string/search/performance/SubstringSearchPerformanceComparison.java
@@ -1,4 +1,4 @@
-package com.baeldung.string.searching;
+package com.baeldung.string.search.performance;
 
 import java.util.concurrent.TimeUnit;
 import java.util.regex.Pattern;
